WIP Update test dir structure
