+++ test/unit/server/services/procedures/chart/xTicks-isTimeSeries.png
