+++ test/unit/server/services/procedures/cloud-variables.spec.js
@@ -3,7 +3,7 @@ describe('cloud-variables', function() {
 
     const Q = require('q');
     const utils = require('../../../../assets/utils');
-    const CloudVariables = utils.reqSrc('rpc/procedures/cloud-variables/cloud-variables');
+    const CloudVariables = utils.reqSrc('services/procedures/cloud-variables/cloud-variables');
     const RPCMock = require('../../../../assets/mock-rpc');
     const cloudvariables = new RPCMock(CloudVariables);
     const assert = require('assert');
