+++ test/unit/server/services/procedures/google-streetview.js
