+++ test/unit/server/services/procedures/star-map.spec.js
